This is my commit, I updated Gillian's code in my Android Studio project, the code works fine and the app runs on my phone. You might encounter some errors feel free to do a pull request to this code also.
